vm/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
requests
schedule
numpy
docker
scikit-learn
